scraper/pyproject.toml
[project]
name = "marginiq"
version = "0.1.0"
description = "Weekly CyberPayOnline scraper for Davis Delivery margin analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pydantic>=2.6",
    "typer>=0.12",
    "python-dotenv>=1.0",
    "firebase-admin>=6.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.20",
]

[project.scripts]
marginiq = "marginiq.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["cyberpay*", "marginiq*"]

[tool.pytest.ini_options]
markers = [
    "live: tests that hit phoenix.cyberpayonline.com (need CYBERPAY_USER/PASS env)",
]
